Grey background in tour chart for night sections, bugfix for tour start components at import and library updates (#52)

+++ bundles/net.tourbook/src/net/tourbook/ui/tourChart/ValuePoint_ToolTip_UI.java
@@ -1706,8 +1706,7 @@ public class ValuePoint_ToolTip_UI extends Pinned_ToolTip_Shell implements IPinn
       }
 
       if (_isVisible_And_Available_TimeOfDay) {
-         _lblTimeOfDay.setText(UI.format_hhh_mm_ss(
-               (_tourData.getStartTimeOfDay() + (long) timeSerie[valueIndex]) % 86400));
+         _lblTimeOfDay.setText(UI.format_hhh_mm_ss((_tourData.getStartTimeOfDay() + timeSerie[valueIndex]) % 86400));
       }
 
       if (_isVisible_And_Available_TimeSlice) {
